+++ backend/libraries/utils/src/time.rs
@@ -9,9 +9,9 @@ pub const WEEK_IN_MS: u64 = DAY_IN_MS * 7;
 const NANOS_PER_MILLISECOND: u64 = 1_000_000;
 
 pub fn now_millis() -> TimestampMillis {
-    ic_cdk::api::time() as u64 / NANOS_PER_MILLISECOND
+    ic_cdk::api::time() / NANOS_PER_MILLISECOND
 }
 
 pub fn now_nanos() -> TimestampNanos {
-    ic_cdk::api::time() as u64
+    ic_cdk::api::time()
 }
